refactor: 재사용 컴포넌트로 수정
